make jaeger compatible with GlobalTracer stuff
